cleanup and notebook
